Add some detail about --bind option in README and CLI printout
I was happy to find out about the `--bind` option when reading over the code, but could not really find it documented anywhere.

Tried to incorporate it into the CLI "help printout" in a better way but had no luck - seems it doesn't have support for e.g. printing a command-specific option alongside the command.

May be worth adding the binding setting to kvass config and set it once via `kvass config bind 1.2.3.4:56`. Documenting that via CLI would go a bit better then, I think.

Please feel free to ignore this PR or change things about it if you want to!

This is a great little tool.